Replace create / update with upsert
This patch replaces both methods with a single method upsert. The idea
behind it, as discussed, is to have a single method for handling queue
creation and update.

The method is expected to return Ture when a queue was created and False
when a queue was updated.

Implements blueprint storage-base
Fixes bug 1155657

Change-Id: I367c5a29038acc89dbf60a3df9d8dcbf436037a7