ClaimMessage implementation for mongodb
This is a first proposal os ClaimMessage support for the mongodb
storage. Please read comments in the code for a more detailed
explanation of this implementation.

Some issues and race conditions might be found and this implementation
tries to cover them in the best way.

More detailed review is required in the claim creation method

Implements blueprint storage-mongodb

Change-Id: I532fdad895a18dc4961a86adbe7086e9eaa231f9