refactor: Hoist helpers.expect into package namespace
It's silly to have to import a package's submodule directly,
since in this case that only introduces another module name
that the developer has to keep track of, just to gain access
to a solitary function. This patch hoists
marconi.tests.util.helpers.expect into marocni.tests.util so
that it can be referenced more naturally by the test author.

Accordingly, the Hacking.rst guide has been updated to allow
this sort of thing (with some discretion).

Implements: blueprint grizzly-debt
Change-Id: Iad2e4f728aa4826d99bec2c599e2947a87eb5945