Pass content length to req.stream.read() in WSGI transport
This update modifies the WSGI transport driver so that it
always passes in a length when reading from the request
stream. This is necessary because WSGI servers are not
required to provide a stream that supports "read all"
semantics. For example, wsgiref will simply block if
read() is called without arguments.

Fixes: bug #1189648
Change-Id: I267e3bce36f0e11d931d9e9800bf3ad3aca7a804