Remove project ID from the URI
This patch modifies the WSGI routes to pull the project ID from
the headers (X-Project-ID) rather than from the URI. Once this
is merged, the implementation will be fully compliant with
the V1 API spec.

Note that clients will be required to include X-Project-ID in
requests since there will not always be a 1:1 relationship
between auth token and project. To support this scenario, the
service provider will also need to perform policy enforcement
in a proxy or in middleware.

Implements: blueprint v1-api
Change-Id: Ib0b1218a98521b903643f3446d2192cddde135a7