Implement grace period for claimed message lifetimes
This commit adds a grace period to claims such that claimed
messages are guaranteed not to expire before the claim
itself expires, plus an additional "grace" number of
seconds.

This provides fudge room so that when one worker crashes,
an alternate worker has some time to reclaim the orphaned
messages once they are released from the stalled claim.

Change-Id: Id0e194a85e67f20babfb97e441e11703d6bf5563
Implements: blueprint claimed-message-grace