Implement v1 JSON Home Document
This patch introduces a JSON Home Document that enables users and api
clients to "follow their nose" in interacting with the Marconi
HTTP API. This is meant to be a baseline implementation, and as such
simply returns a hard-coded document.

Future iterations may include support for vendor-specific JSON and
dynamic, RBAC-aware home documents.

This patch anticipates the creation of http://docs.openstack-marconi.org

Implements: blueprint v1-api
Change-Id: Ie13e77fc220884bbb97aa988ddf2bc85076e37a1