fix(storage/mongo): unclaimed messages not listed
Adds a unit test that reproduces this error and applies the fix.

The fix is subtle. When messages are unclaimed, their claim expiration
time and claim ID are set to 0 and None respectively. The cause for
the bug is that claim expiration time is set to 0, rather than a valid
datetime object. As a result, the comparison silently fails.

The fix: fast-forward the expiration time of a claim to now.

Change-Id: I7781795c394b2a471012e072a522ee228d56a894
Closes-Bug: #1210633