Fixes unique constraint regex
This patch fixes the regex being used to extract the marker from the
unique index error. The issue was caused because we introduced project
to the queue_marker index and it wasn't being considered in that regex.

This regex is more general. It matches the index name and gets the
marker from the end of the error message.

Change-Id: I02bc153636f10bc49b8634fc4bfeb4bf2b6d6597
Closes-bug: #1211386