Don't wait for ack when deleting expired messages
This patch simply sets write concern to 0 in the MongoDB driver,
when removing expired messages. This shortens the amount of time
that garbage collection takes, and slightly reduces the amount
of traffic going between app and db servers, as well as between
master and secondary nodes.

This is a safe thing to do, since the operator will be running
multiple GC daemons, so that in the case of a failure, the
operation will be retried in the near future.

The downside is that persistent failures to collect expired
messages will cause the DB to grow out of control; however,
it is expected that if this operation is failing, more than
likely message CRUD in general is failing, which mitigates
the risk just mentioned.

Change-Id: I90950721361dacffa508624281d2e2ea8cf53724