Use oslo.config for functional tests
Instead of defining a Config class for functional tests, this patch
integrates it with Marconi's base test suite and replaces the use of
ConfigParser with olso.config.

The patch doesn't change current behavior and gives the possibility to
use external configuration files for both the functional test and
Marconi.

blueprint refactor-system-tests

Change-Id: I60de6e199182881e945a24525c7c4f2d6937256f