fix: Claim can return 404
This patch changes the way creating a claim behaves when the
queue does not exist. Instead of returning 404, it returns
204, meaning the claim was not able to find any messages to
claim.

This was done for two reasons:

1. For eventually-consistent backends, a brand new queue may
not appear to exist, so this new semantic allows the driver
to just try to grab some messages without checking first
whether the queue "exists".
2. For backends like MongoDB that require an extra check to
determine whether a queue exists, this removes an extra
round trip to the DB in order to perform the operation.

Note that the SQLite driver was updated in order to be
consistent with the new behavior.

Change-Id: Icde4ce493a76ef145e99300b67d8344d5092e38d
Partial-Bug: #1218990