Make tox use develop instead of sdist
tox 1.6 was released, which means that we can now take advantage of the
feature we added to it - which is using setup.py develop to install the
code into the virtualenv. The logic was taken from run_tests.sh - so the
performance issues around using tox vs. using install_venv should now be
gone.

This patch requires tox >=1.6.

Change-Id: I505fb530e8047f478e98ce2fd6c01776b147d87f