feat: define interfaces for proxy storage
This patchset introduces an interface for future proxy storage drivers
to work with. Along with the interface, the basic machinery for
writing unit tests for the storage drivers is exposed.

It introduces two controllers:
- PartitionsController
- CatalogueController

This should facilitate exposing stevedore entry points for defining
custom proxy storage drivers.

Appropriate exception classes are also exposed.

six is introduced as a dependency to take advantage of
six.add_metaclass.

Change-Id: I49f66a942dc24ef3e090132c6862529fa4ebfcec