perf(mongodb): Combine project and queue message fields
This patch merges the message and queue collections' project name and
queue name fields into a single 'p_q' field, named to denote the
compound nature of the scoped queue name string.

This should reduce index complexity, making for faster inserts and
finds. This hypothesis obviously needs to be verified during the
review process for this patch.

Change-Id: I3da09806ad50c2d79f1f0c03bee68f1cc4ee787d
Implements: blueprint api-perf-tuning