fix(mongodb): Remove $or clause used to claim messages
The $or clause was removed from one of the queries in
ClaimController.create() because it is no longer necessary,
and makes the query slower than it would otherwise be.

Mongo will just use the _id_ index now, which should be fine
since scanning a small set of messages by c.e is fast.

Change-Id: Ifcb918ef66d072a10690ad49406f4a787e84becd
Implements: blueprint bp/api-perf-tuning