fix(wsgi): Non-ASCII characters in the queue name trigger 500 status
The problem was that queue_name wasn't being validated before being
used to construct log messages. Also, I found that the regex used to
validate queue names was not reliable and needed to be rewritten.

Upon further investigation, I also discovered that the queue name
wasn't being validated for all routes that included it. This was
also fixed. Validating all input is a basic best practice re
application security.

Along the way, I edited the validation error messages for clarity,
and relocated the ValidationFailed exception since it is only used
by the transport driver.

Change-Id: Idf9f6c7b9a88865a6396dfcd72abbce96c774490
Closes-Bug: 1208873