fix(mongo): Queue listing may include queues from other projects
This patch changes the mongo driver so that it explicitly checks that the
queue belongs to the correct project, rather than simply getting
everything greater than a specific queue name.

The associated unit test was updated to catch any regressions in the
fix. The test was also updated to take into account the change in queue
metadata handling that landed several weeks ago.

Change-Id: I6434d8452ff38bb3191ab947f5f143b650e2b103
Closes-Bug: #1236605