feat: add shard management resource
This patchset adds a new resource that exposes shard registry through
the administrative API. Since there is no storage backend available
that implements this functionality yet, the shards resource is not
connected to the queues routes just yet.

Preliminary support is added for unit testing this soon.

Change-Id: I49793990327643fc2bc5091615fb0bebcef04bc0
Partially-implements: blueprint storage-sharding