fix: mongodb tests consume storage like crazy
This patchset addresses a problem where _purge_databases wasn't being
called in the storage tests for mongodb. This was consuming about
~19GB of storage per full suite run.

The problem was caused by _purge_databases not being overridden by the
MongodbMixin class. It was a case of the diamond problem in an
inheritance graph. Here's the summary:

* testing.TestBase defines _purge_databases
* ControllerTestBase is a child of testing.TestBase
* XControllerTest are children of ControllerTestBase
* MongodbTestMixin was a child of object, and reimplements
  _purge_databases
* When MongoXControllerTest was presented with parents:
    - XControllerTest
    - MongodbTestMixin

  The MRO algorithm chose the implementation of _purge_databases that
  was first presented, in this case, the null implementation provided
  by testing.TestBase

The proposed solution for the time being is to have each of the
MongoXControllerTest classes to override the _purge_databases method.

Note: subtyping can get very messy.

Change-Id: Ia1abc4c6ec682f3d176f2093d45139f45f7ac12a