Enforce size verification on content-length
Closes-Bug: #1261579

Previously, Marconi used to enforce the message and queue size limit on
the resource itself. In the case of a Message, it'd verify that the
message body wouldn't exceed the message size limit. This enforcement
has been moved, by this patch, up to the request size.

For both cases, the content_length is required to be smaller than the
size limit configured by the service admin. None of the default values
were modified.

Change-Id: I06e28acaac234242a40cd2afafb12be067851c20