feat(v1.1): Plumb v1.1 API in the WSGI transport
This patch creates a copy of the v1.0 API as a starting point for
implementing the v1.1 API. The way that the "admin_mode" config option was
implemented was changed, in order to simplify the grafting of the two API
versions into a single driver. When the mode was implemented via inheritance,
grafting the two API versions required some ugly hacks.

Partially-Implements: blueprint api-v1.1
Change-Id: I0e753315e3171bb5c23a57b8fb3826d08bfd5f26